Dynamically load brick UI
