+++ src/bricks/renderers/propertyTable.tsx
@@ -31,6 +31,7 @@ interface Item {
   children: Item[];
 }
 
+// FIXME: does this need to get extracted to a different file to avoid the react import?
 function richValue(value: unknown): unknown {
   if (typeof value === "string" && isValidUrl(value)) {
     return (
